Extract part of the BackupDelta component "MetaDiff"
- It's for easily to insert it into the Backup Status block
- Show BackupDelta only if it has realtime backups
