+++ client/landing/jetpack-cloud/components/backup-delta/style.scss
@@ -1,5 +1,5 @@
-.backup-delta__view-all-button {
-    float: none;
+.backup-delta {
+	margin: 0 16px;
 }
 
 .backup-delta__realtime .activity-log-item__actor {
@@ -9,124 +9,6 @@
 .backup-delta__realtime .button.is-borderless.is-primary {
     float: none;
     display: inline-block;
-}
-
-.backup-delta__changes-header {
-    margin-top: 32px;
-	font-size: 16px;
-    font-weight: 600;
-	border-bottom: 1px solid var( --studio-gray-10 );
-	padding-bottom: 16px;
-}
-
-.backup-delta__section-header {
-    margin: 0.5rem 0;
-
-}
-
-.backup-delta__section-posts {
-    margin-bottom: 1rem;
-
-}
-
-.backup-delta__media-image {
-    display: inline-block;
-    margin-right: 1rem;
-    position: relative;
-}
-
-.backup-delta__media-image img {
-    object-fit: cover;
-    width: 6rem;
-    height: 6rem;
-}
-
-.backup-delta__media-title {
-    position: relative;
-    top: -1.9rem;
-    background-color: rgba( 0, 0, 0, 0.5 );
-    color: #fff;
-    font-size: 0.8rem;
-    padding-top: 0.3rem;
-}
-
-.backup-delta__media-title .gridicon {
-    width: 0.7rem;
-    height: 1rem;
-    margin: 0 0.3rem;
-}
-
-.backup-delta__media-title-text {
-    position: relative;
-    top: -0.18rem;
-    display: inline-block;
-}
-
-.backup-delta__count-bubble {
-    display: inline-block;
-    background: #dcdcde;
-    color: #000;
-    border-radius: 1rem;
-    padding: 0.2rem 0.5rem;
-}
-
-.backup-delta__deleted-count-bubble {
-    position: absolute;
-    right: 0.5rem;
-    top: 2rem;
-    display: inline-block;
-    background: #d63638;
-    color: #fff;
-    border-radius: 1rem;
-    padding: 0.2rem 0.5rem;
-}
-
-.gridicon.backup-delta__post-icon {
-    width: 1rem;
-    fill: #787c82;
-}
-
-.gridicon.backup-delta__theme-icon-installed {
-    width: 1rem;
-    height: 1rem;
-    fill: #349e0b;
-}
-
-.gridicon.backup-delta__theme-icon-removed {
-    width: 1rem;
-    height: 1rem;
-    fill: #d63638;
-}
-
-.backup-delta__extension-block-installed {
-    color: #349e0b;
-}
-
-.backup-delta__extension-block-removed {
-    color: #d63638;
-}
-
-.backup-delta__post-link {
-    display: inline-block;
-    position: relative;
-    top: -0.4rem;
-    left: 0.5rem;
-    color: #349e0b;
-}
-
-.backup-delta__post-block {
-    margin-bottom: -0.5rem;
-}
-
-.backup-delta__metas {
-    margin: 1rem 0;
-}
-
-.backup-delta__extension-block-text {
-    display: inline-block;
-    position: relative;
-    left: 0.2rem;
-    top: -0.1rem;
 }
 
 .backup-delta__realtime-cards {
@@ -141,29 +23,4 @@
 
 .backup-delta__realtime-emptyday {
     margin-top: 1rem;
-}
-
-.backup-delta__changes-header:first-child {
-    margin-top: 0;
-}
-
-.backup-delta__daily-no-changes {
-    font-style: italic;
-	padding-top: 16px;
-	color: var( --studio-gray-40 );
-}
-
-.backup-delta__daily {
-	margin: 32px -16px 0;
-	padding: 0 16px;
-	text-align: left;
-}
-
-.backup-delta__view-all-button {
-	display: block;
-	text-align: center;
-}
-
-.backup-delta__metas {
-	margin-top: 0;
 }
