+++ client/landing/jetpack-cloud/components/daily-backup-status/style.scss
@@ -153,6 +153,154 @@
 	margin-bottom: 1.5rem;
 }
 
+.daily-backup-status__view-all-button {
+	float: none;
+}
+
+.daily-backup-status__changes-header {
+	margin-top: 32px;
+	font-size: 16px;
+	font-weight: 600;
+	border-bottom: 1px solid var( --studio-gray-10 );
+	padding-bottom: 16px;
+}
+
+.daily-backup-status__section-header {
+	margin: 0.5rem 0;
+
+}
+
+.daily-backup-status__section-posts {
+	margin-bottom: 1rem;
+
+}
+
+.daily-backup-status__media-image {
+	display: inline-block;
+	margin-right: 1rem;
+	position: relative;
+}
+
+.daily-backup-status__media-image img {
+	object-fit: cover;
+	width: 6rem;
+	height: 6rem;
+}
+
+.daily-backup-status__media-title {
+	position: relative;
+	top: -1.9rem;
+	background-color: rgba( 0, 0, 0, 0.5 );
+	color: #fff;
+	font-size: 0.8rem;
+	padding-top: 0.3rem;
+}
+
+.daily-backup-status__media-title .gridicon {
+	width: 0.7rem;
+	height: 1rem;
+	margin: 0 0.3rem;
+}
+
+.daily-backup-status__media-title-text {
+	position: relative;
+	top: -0.18rem;
+	display: inline-block;
+}
+
+.daily-backup-status__count-bubble {
+	display: inline-block;
+	background: #dcdcde;
+	color: #000;
+	border-radius: 1rem;
+	padding: 0.2rem 0.5rem;
+}
+
+.daily-backup-status__deleted-count-bubble {
+	position: absolute;
+	right: 0.5rem;
+	top: 2rem;
+	display: inline-block;
+	background: #d63638;
+	color: #fff;
+	border-radius: 1rem;
+	padding: 0.2rem 0.5rem;
+}
+
+.gridicon.daily-backup-status__post-icon {
+	width: 1rem;
+	fill: #787c82;
+}
+
+.gridicon.daily-backup-status__theme-icon-installed {
+	width: 1rem;
+	height: 1rem;
+	fill: #349e0b;
+}
+
+.gridicon.daily-backup-status__theme-icon-removed {
+	width: 1rem;
+	height: 1rem;
+	fill: #d63638;
+}
+
+.daily-backup-status__extension-block-installed {
+	color: #349e0b;
+}
+
+.daily-backup-status__extension-block-removed {
+	color: #d63638;
+}
+
+.daily-backup-status__post-link {
+	display: inline-block;
+	position: relative;
+	top: -0.4rem;
+	left: 0.5rem;
+	color: #349e0b;
+}
+
+.daily-backup-status__post-block {
+	margin-bottom: -0.5rem;
+}
+
+
+.daily-backup-status__extension-block-text {
+	display: inline-block;
+	position: relative;
+	left: 0.2rem;
+	top: -0.1rem;
+}
+
+.daily-backup-status__metas {
+	margin: 1rem 0;
+}
+
+.daily-backup-status__changes-header:first-child {
+	margin-top: 0;
+}
+
+.daily-backup-status__daily-no-changes {
+	font-style: italic;
+	padding-top: 16px;
+	color: var( --studio-gray-40 );
+}
+
+.daily-backup-status__daily {
+	margin: 32px -16px 0;
+	padding: 0 16px;
+	text-align: left;
+}
+
+.daily-backup-status__view-all-button {
+	display: block;
+	text-align: center;
+}
+
+.daily-backup-status__metas {
+	margin-top: 0;
+}
+
 @include breakpoint( '>660px' ) {
 	.daily-backup-status__meta,
 	.daily-backup-status__title {
@@ -203,7 +351,7 @@
 		padding-left: 2rem;
 		padding-right: 2rem;
 	}
-	
+
 	.daily-backup-status__failed-message {
 		font-size: 36px;
 		font-weight: 600;
